refactored options, extensions, partitions
