+++ src/app/(protected)/quote/ClientQuote.js
@@ -1,11 +1,9 @@
 'use client';
-import { useState, useEffect, Fragment } from 'react';
-import Link from 'next/link';
+import { useState, useEffect, Fragment, act } from 'react';
 import Image from 'next/image';
 import styles from './page.module.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
-  faHouse,
   faChevronLeft,
   faChevronRight,
   faTrash,
@@ -13,7 +11,7 @@ import {
   faPlus,
   faCheck,
 } from '@fortawesome/free-solid-svg-icons';
-// import { useFormState, useNavigation } from '../../../hooks';
+
 import useFormState from '../../../hooks/useFormState';
 import useNavigation from '../../../hooks/useNavigation';
 
@@ -22,38 +20,16 @@ import { initialState } from './_initialState';
 import QuoteInformation from '../../../components/quoteSections/QuoteInformation';
 import DesignCodes from '../../../components/quoteSections/DesignCodes';
 import BuildingLayout from '../../../components/quoteSections/BuildingLayout';
+import BuildingOptions from '../../../components/quoteSections/BuildingOptions';
+import BuildingExtensions from '../../..//components/quoteSections/BuildingExtensions';
+import BuildingPartitions from '../../..//components/quoteSections/BuildingPartitions';
 
 import CopyBuildingDialog from '../../../components/CopyBuildingDialog';
 import DeleteDialog from '../../../components/DeleteDialog';
 import ReusableSelect from '../../../components/ReusableSelect';
 import BuildingSketch from '../../../components/BuildingSketch';
 import { logo } from '../../../../public/images';
-import {
-  shapes,
-  frames,
-  FrameOptions,
-  SidewallBracingType,
-  EndwallBracingType,
-  breakPoints,
-  girtTypes,
-  girtSpacing,
-  baseCondition,
-  purlinSpacing,
-  roofPanels,
-  roofGauge,
-  roofFinish,
-  wallPanels,
-  wallGauge,
-  wallFinish,
-  soffitPanels,
-  soffitGauge,
-  soffitFinish,
-  roofInsulation,
-  wallInsulation,
-  extInsulation,
-  orientations,
-  walls,
-} from '../../../util/dropdownOptions';
+
 import PageHeader from '@/components/PageHeader';
 
 export default function ClientQuote({ session }) {
@@ -62,12 +38,10 @@ export default function ClientQuote({ session }) {
   const [activeBuilding, setActiveBuilding] = useState(0);
   const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
   const [buildingToDelete, setBuildingToDelete] = useState(null);
-  const [activeCanopy, setActiveCanopy] = useState(0);
-  const [activePartition, setActivePartition] = useState(0);
-  const [activeLinerPanel, setActiveLinerPanel] = useState(0);
 
   const [dialogOpen, setDialogOpen] = useState(false);
   const [sourceBuildingIndex, setSourceBuildingIndex] = useState(0);
+
   // Hooks
   const {
     values,
@@ -77,6 +51,9 @@ export default function ClientQuote({ session }) {
     handleCanopyChange,
     handlePartitionChange,
     handleLinerPanelChange,
+    handleWainscotChange,
+    handlePartialWallChange,
+    handleWallSkirtChange,
     setValues,
   } = useFormState(initialState);
 
@@ -176,95 +153,29 @@ export default function ClientQuote({ session }) {
           canopies: [],
           partitions: [],
           linerPanels: [],
+          wainscots: [],
+          partialWalls: [],
+          wallSkirts: [],
+          fswPolySize: '3',
+          fswPolyColor: 'clear',
+          fswPolyQty: '',
+          bswPolySize: '3',
+          bswPolyColor: 'clear',
+          bswPolyQty: '',
+          lewPolySize: '3',
+          lewPolyColor: 'clear',
+          lewPolyQty: '',
+          rewPolySize: '3',
+          rewPolyColor: 'clear',
+          rewPolyQty: '',
+          backRoofPolySize: '10',
+          backRoofPolyColor: 'clear',
+          backRoofPolyQty: '',
+          frontRoofPolySize: '10',
+          frontRoofPolyColor: 'clear',
+          frontRoofPolyQty: '',
         },
       ],
-    }));
-  };
-
-  const addCanopy = (buildingIndex) => {
-    setValues((prev) => ({
-      ...prev,
-      buildings: prev.buildings.map((building, index) =>
-        index === buildingIndex
-          ? {
-              ...building,
-              canopies: [
-                ...building.canopies,
-                {
-                  wall: 'frontSidewall',
-                  width: '',
-                  slope: '',
-                  startBay: '',
-                  endBay: '',
-                  elevation: '',
-                  addColumns: false,
-                  roofPanelType: 'pbr',
-                  roofPanelGauge: '',
-                  roofPanelFinish: '',
-                  soffitPanelType: 'tuff',
-                  soffitPanelGauge: '',
-                  soffitPanelFinish: '',
-                },
-              ],
-            }
-          : building
-      ),
-    }));
-  };
-
-  const addPartition = (buildingIndex) => {
-    setValues((prev) => ({
-      ...prev,
-      buildings: prev.buildings.map((building, index) =>
-        index === buildingIndex
-          ? {
-              ...building,
-              partitions: [
-                ...building.partitions,
-                {
-                  orientation: 't',
-                  start: '',
-                  end: '',
-                  offset: '',
-                  height: '',
-                  baySpacing: '',
-                  insulation: 'vrr4',
-                  leftPanelType: 'pbr',
-                  leftPanelGauge: '',
-                  leftPanelFinish: '',
-                  rightPanelType: 'pbr',
-                  rightPanelGauge: '',
-                  rightPanelFinish: '',
-                },
-              ],
-            }
-          : building
-      ),
-    }));
-  };
-
-  const addLinerPanel = (buildingIndex) => {
-    setValues((prev) => ({
-      ...prev,
-      buildings: prev.buildings.map((building, index) =>
-        index === buildingIndex
-          ? {
-              ...building,
-              linerPanels: [
-                ...building.linerPanels,
-                {
-                  wall: 'frontSidewall',
-                  start: '',
-                  end: '',
-                  height: '',
-                  panelType: 'pbr',
-                  panelGauge: '',
-                  panelFinish: '',
-                },
-              ],
-            }
-          : building
-      ),
     }));
   };
 
@@ -281,80 +192,6 @@ export default function ClientQuote({ session }) {
       // If a building before the active one is removed, adjust the active index
       setActiveBuilding((prev) => prev - 1);
     }
-  };
-
-  const removeCanopy = (buildingIndex, canopyIndex) => {
-    setValues((prev) => {
-      const newBuildings = prev.buildings.map((building, bIndex) =>
-        bIndex === buildingIndex
-          ? {
-              ...building,
-              canopies: building.canopies.filter(
-                (_, cIndex) => cIndex !== canopyIndex
-              ),
-            }
-          : building
-      );
-
-      // Update activeCanopy if necessary
-      const remainingCanopies = newBuildings[buildingIndex].canopies.length;
-      if (canopyIndex <= activeCanopy && activeCanopy > 0) {
-        setActiveCanopy(Math.min(activeCanopy - 1, remainingCanopies - 1));
-      }
-
-      return { ...prev, buildings: newBuildings };
-    });
-  };
-
-  const removePartition = (buildingIndex, partitionIndex) => {
-    setValues((prev) => {
-      const newBuildings = prev.buildings.map((building, bIndex) =>
-        bIndex === buildingIndex
-          ? {
-              ...building,
-              partitions: building.partitions.filter(
-                (_, pIndex) => pIndex !== partitionIndex
-              ),
-            }
-          : building
-      );
-
-      // Update activePartition if necessary
-      const remainingPartitions = newBuildings[buildingIndex].partitions.length;
-      if (partitionIndex <= activePartition && activePartition > 0) {
-        setActivePartition(
-          Math.min(activePartition - 1, remainingPartitions - 1)
-        );
-      }
-
-      return { ...prev, buildings: newBuildings };
-    });
-  };
-
-  const removeLinerPanel = (buildingIndex, linerPanelIndex) => {
-    setValues((prev) => {
-      const newBuildings = prev.buildings.map((building, bIndex) =>
-        bIndex === buildingIndex
-          ? {
-              ...building,
-              linerPanels: building.linerPanels.filter(
-                (_, lpIndex) => lpIndex !== linerPanelIndex
-              ),
-            }
-          : building
-      );
-
-      // Update activePartition if necessary
-      const remainingLinerPanels =
-        newBuildings[buildingIndex].linerPanels.length;
-      if (linerPanelIndex <= activeLinerPanel && activeLinerPanel > 0) {
-        setActiveLinerPanel(
-          Math.min(activeLinerPanel - 1, remainingLinerPanels - 1)
-        );
-      }
-
-      return { ...prev, buildings: newBuildings };
-    });
   };
 
   const openCopyDialog = (index) => {
@@ -430,42 +267,6 @@ export default function ClientQuote({ session }) {
     // Here you would typically send the data to your backend
   };
 
-  const selectedSoffitPanel = soffitPanels.find(
-    (panel) => panel.id === values.buildings[activeBuilding].soffitPanelType
-  );
-
-  const selectedCanopyRoofPanel = roofPanels.find(
-    (panel) =>
-      panel.id ===
-      values.buildings[activeBuilding].canopies[activeCanopy]?.roofPanelType
-  );
-
-  const selectedCanopySoffitPanel = soffitPanels.find(
-    (panel) =>
-      panel.id ===
-      values.buildings[activeBuilding].canopies[activeCanopy]?.soffitPanelType
-  );
-
-  const selectedPartitionLeftPanel = wallPanels.find(
-    (panel) =>
-      panel.id ===
-      values.buildings[activeBuilding].partitions[activePartition]
-        ?.leftPanelType
-  );
-
-  const selectedPartitionRightPanel = wallPanels.find(
-    (panel) =>
-      panel.id ===
-      values.buildings[activeBuilding].partitions[activePartition]
-        ?.rightPanelType
-  );
-
-  const selectedLinerPanel = wallPanels.find(
-    (panel) =>
-      panel.id ===
-      values.buildings[activeBuilding].linerPanels[activeLinerPanel]?.panelType
-  );
-
   // Checking for screen width to conditionally render DOM elements
   useEffect(() => {
     if (window.innerWidth > 1000) {
@@ -488,7 +289,7 @@ export default function ClientQuote({ session }) {
   return (
     <main>
       <PageHeader session={session} title="Quote Input" isLogOut={false} />
-      {/* <div> */}
+      {/* Sidebar Navigation */}
       {isDesktop && (
         <div>
           <nav className={styles.sidebar}>
@@ -513,9 +314,9 @@ export default function ClientQuote({ session }) {
             <button onClick={() => setActiveCardDirectly('bldg-options')}>
               Building {activeBuilding + 1} - Options
             </button>
-            <button onClick={() => setActiveCardDirectly('bldg-cranes')}>
+            {/* <button onClick={() => setActiveCardDirectly('bldg-cranes')}>
               Building {activeBuilding + 1} - Cranes
-            </button>
+            </button> */}
             <button onClick={() => setActiveCardDirectly('bldg-openings')}>
               Building {activeBuilding + 1} - Openings
             </button>
@@ -713,1236 +514,42 @@ export default function ClientQuote({ session }) {
         {/* Building Extensions Page */}
         {activeCard == 'bldg-extensions' && (
           <>
-            <section className="card">
-              <header>
-                <h3>Roof Extensions</h3>
-              </header>
-              <div className="cardGrid">
-                <div className="cardInput">
-                  <label
-                    htmlFor={`buildingFswExtensionWidth-${activeBuilding}`}
-                  >
-                    Front Sidewall Extension Width:
-                  </label>
-                  <input
-                    type="text"
-                    id={`buildingFswExtensionWidth-${activeBuilding}`}
-                    name={`buildingFswExtensionWidth-${activeBuilding}`}
-                    value={values.buildings[activeBuilding].fswExtensionWidth}
-                    onChange={(e) =>
-                      handleNestedChange(
-                        activeBuilding,
-                        'fswExtensionWidth',
-                        e.target.value
-                      )
-                    }
-                    placeholder="Feet"
-                  />
-                </div>
-                <div className="cardInput">
-                  <label
-                    htmlFor={`buildingBswExtensionWidth-${activeBuilding}`}
-                  >
-                    Back Sidewall Extension Width:
-                  </label>
-                  <input
-                    type="text"
-                    id={`buildingBswExtensionWidth-${activeBuilding}`}
-                    name={`buildingBswExtensionWidth-${activeBuilding}`}
-                    value={values.buildings[activeBuilding].bswExtensionWidth}
-                    onChange={(e) =>
-                      handleNestedChange(
-                        activeBuilding,
-                        'bswExtensionWidth',
-                        e.target.value
-                      )
-                    }
-                    placeholder="Feet"
-                  />
-                </div>
-                <div className="cardInput">
-                  <label
-                    htmlFor={`buildingLewExtensionWidth-${activeBuilding}`}
-                  >
-                    Left Endwall Extension Width:
-                  </label>
-                  <input
-                    type="text"
-                    id={`buildingLewExtensionWidth-${activeBuilding}`}
-                    name={`buildingLewExtensionWidth-${activeBuilding}`}
-                    value={values.buildings[activeBuilding].lewExtensionWidth}
-                    onChange={(e) =>
-                      handleNestedChange(
-                        activeBuilding,
-                        'lewExtensionWidth',
-                        e.target.value
-                      )
-                    }
-                    placeholder="Feet"
-                  />
-                </div>
-                <div className="cardInput">
-                  <label
-                    htmlFor={`buildingRewExtensionWidth-${activeBuilding}`}
-                  >
-                    Right Endwall Extension Width
-                  </label>
-                  <input
-                    type="text"
-                    id={`buildingRewExtensionWidth-${activeBuilding}`}
-                    name={`buildingRewExtensionWidth-${activeBuilding}`}
-                    value={values.buildings[activeBuilding].rewExtensionWidth}
-                    onChange={(e) =>
-                      handleNestedChange(
-                        activeBuilding,
-                        'rewExtensionWidth',
-                        e.target.value
-                      )
-                    }
-                    placeholder="Feet"
-                  />
-                </div>
-                <div className="cardInput">
-                  <label
-                    htmlFor={`buildingFrontExtensionBays-${activeBuilding}`}
-                  >
-                    Front Extension Bays:
-                  </label>
-                  <input
-                    type="text"
-                    id={`buildingFrontExtensionBays-${activeBuilding}`}
-                    name={`buildingFrontExtensionBays-${activeBuilding}`}
-                    value={values.buildings[activeBuilding].frontExtensionBays}
-                    onChange={(e) =>
-                      handleNestedChange(
-                        activeBuilding,
-                        'frontExtensionBays',
-                        e.target.value
-                      )
-                    }
-                    placeholder="Separate Bays with Space"
-                  />
-                </div>
-                <div className="checkRow">
-                  <input
-                    type="checkbox"
-                    id={`buildingFrontExtensionColumns-${activeBuilding}`}
-                    name={`buildingFrontExtensionColumns-${activeBuilding}`}
-                    checked={
-                      values.buildings[activeBuilding].frontExtensionColumns
-                    }
-                    onChange={(e) =>
-                      handleNestedChange(
-                        activeBuilding,
-                        'frontExtensionColumns',
-                        e.target.checked
-                      )
-                    }
-                  />
-                  <label
-                    htmlFor={`buildingFrontExtensionColumns-${activeBuilding}`}
-                  >
-                    Add Columns
-                  </label>
-                </div>
-                <div className="cardInput">
-                  <label
-                    htmlFor={`buildingBackExtensionBays-${activeBuilding}`}
-                  >
-                    Back Extension Bays:
-                  </label>
-                  <input
-                    type="text"
-                    id={`buildingBackExtensionBays-${activeBuilding}`}
-                    name={`buildingBackExtensionBays-${activeBuilding}`}
-                    value={values.buildings[activeBuilding].backExtensionBays}
-                    onChange={(e) =>
-                      handleNestedChange(
-                        activeBuilding,
-                        'backExtensionBays',
-                        e.target.value
-                      )
-                    }
-                    placeholder="Separate Bays with Space"
-                  />
-                </div>
-                <div className="checkRow">
-                  <input
-                    type="checkbox"
-                    id={`buildingBackExtensionColumns-${activeBuilding}`}
-                    name={`buildingBackExtensionColumns-${activeBuilding}`}
-                    checked={
-                      values.buildings[activeBuilding].backExtensionColumns
-                    }
-                    onChange={(e) =>
-                      handleNestedChange(
-                        activeBuilding,
-                        'backExtensionColumns',
-                        e.target.checked
-                      )
-                    }
-                  />
-                  <label
-                    htmlFor={`buildingBackExtensionColumns-${activeBuilding}`}
-                  >
-                    Add Columns
-                  </label>
-                </div>
-              </div>
-              <div className="divider"></div>
-
-              <div className="extendGrid">
-                <div className="extGrid start">
-                  <div className="cardInput">
-                    <ReusableSelect
-                      id={`buildingExtensionInsulation-${activeBuilding}`}
-                      name={`buildingExtensionInsulation-${activeBuilding}`}
-                      value={
-                        values.buildings[activeBuilding].extensionInsulation
-                      }
-                      onChange={(e) =>
-                        handleNestedChange(
-                          activeBuilding,
-                          'extensionInsulation',
-                          e.target.value
-                        )
-                      }
-                      options={extInsulation}
-                      label="Insulation In Extension:"
-                    />
-                  </div>
-                </div>
-
-                <div className="extGrid start">
-                  <div className="cardInput">
-                    <ReusableSelect
-                      id={`buildingSoffitPanels-${activeBuilding}`}
-                      name={`buildingSoffitPanels-${activeBuilding}`}
-                      value={values.buildings[activeBuilding].soffitPanelType}
-                      onChange={(e) =>
-                        handleNestedChange(
-                          activeBuilding,
-                          'soffitPanelType',
-                          e.target.value
-                        )
-                      }
-                      options={soffitPanels}
-                      label="Soffit Panels:"
-                    />
-                  </div>
-                  {values.buildings[activeBuilding].soffitPanelType !=
-                    'none' && (
-                    <>
-                      <div className="cardInput">
-                        <ReusableSelect
-                          id={`buildingSoffitGauge-${activeBuilding}`}
-                          name={`buildingSoffitGauge-${activeBuilding}`}
-                          value={
-                            values.buildings[activeBuilding].soffitPanelGauge
-                          }
-                          onChange={(e) =>
-                            handleNestedChange(
-                              activeBuilding,
-                              'soffitPanelGauge',
-                              e.target.value
-                            )
-                          }
-                          options={soffitGauge}
-                          label="Gauge:"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <ReusableSelect
-                          id={`buildingSoffitFinish-${activeBuilding}`}
-                          name={`buildingSoffitFinish-${activeBuilding}`}
-                          value={
-                            values.buildings[activeBuilding].soffitPanelFinish
-                          }
-                          onChange={(e) =>
-                            handleNestedChange(
-                              activeBuilding,
-                              'soffitPanelFinish',
-                              e.target.value
-                            )
-                          }
-                          options={soffitFinish}
-                          label="Finish:"
-                        />
-                      </div>
-                      {selectedSoffitPanel && selectedSoffitPanel.image && (
-                        <Image
-                          alt={`${selectedSoffitPanel.label}`}
-                          src={selectedSoffitPanel.image}
-                          className="panelImage"
-                        />
-                      )}
-                    </>
-                  )}
-                </div>
-              </div>
-            </section>
-
-            <section className="card">
-              <header>
-                <h3>Canopies</h3>
-              </header>
-              <div className="tableGrid">
-                {values.buildings[activeBuilding].canopies.map(
-                  (canopy, canopyIndex) => (
-                    <Fragment
-                      key={`building-${activeBuilding}-canopy-${canopyIndex}`}
-                    >
-                      <div className="cardInput">
-                        <ReusableSelect
-                          id={`building-${activeBuilding}-canopyWall-${canopyIndex}`}
-                          name={`building-${activeBuilding}-canopyWall-${canopyIndex}`}
-                          value={canopy.wall}
-                          onChange={(e) =>
-                            handleCanopyChange(
-                              activeBuilding,
-                              canopyIndex,
-                              'wall',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeCanopy !== canopyIndex) {
-                              setActiveCanopy(canopyIndex);
-                            }
-                          }}
-                          options={walls}
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-canopyWidth-${canopyIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-canopyWidth-${canopyIndex}`}
-                          name={`building-${activeBuilding}-canopyWidth-${canopyIndex}`}
-                          value={canopy.width}
-                          onChange={(e) =>
-                            handleCanopyChange(
-                              activeBuilding,
-                              canopyIndex,
-                              'width',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeCanopy !== canopyIndex) {
-                              setActiveCanopy(canopyIndex);
-                            }
-                          }}
-                          placeholder="Feet"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-canopySlope-${canopyIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-canopySlope-${canopyIndex}`}
-                          name={`building-${activeBuilding}-canopySlope-${canopyIndex}`}
-                          value={canopy.slope}
-                          onChange={(e) =>
-                            handleCanopyChange(
-                              activeBuilding,
-                              canopyIndex,
-                              'slope',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeCanopy !== canopyIndex) {
-                              setActiveCanopy(canopyIndex);
-                            }
-                          }}
-                          placeholder="x:12"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-canopyStartBay-${canopyIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-canopyStartBay-${canopyIndex}`}
-                          name={`building-${activeBuilding}-canopyStartBay-${canopyIndex}`}
-                          value={canopy.startBay}
-                          onChange={(e) =>
-                            handleCanopyChange(
-                              activeBuilding,
-                              canopyIndex,
-                              'startBay',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeCanopy !== canopyIndex) {
-                              setActiveCanopy(canopyIndex);
-                            }
-                          }}
-                          placeholder="Bay #"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-canopyEndBay-${canopyIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-canopyEndBay-${canopyIndex}`}
-                          name={`building-${activeBuilding}-canopyEndBay-${canopyIndex}`}
-                          value={canopy.endBay}
-                          onChange={(e) =>
-                            handleCanopyChange(
-                              activeBuilding,
-                              canopyIndex,
-                              'endBay',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeCanopy !== canopyIndex) {
-                              setActiveCanopy(canopyIndex);
-                            }
-                          }}
-                          placeholder="Bay #"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-canopyElevation-${canopyIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-canopyElevation-${canopyIndex}`}
-                          name={`building-${activeBuilding}-canopyElevation-${canopyIndex}`}
-                          value={canopy.elevation}
-                          onChange={(e) =>
-                            handleCanopyChange(
-                              activeBuilding,
-                              canopyIndex,
-                              'elevation',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeCanopy !== canopyIndex) {
-                              setActiveCanopy(canopyIndex);
-                            }
-                          }}
-                          placeholder="Feet"
-                        />
-                      </div>
-                      <div className="checkRow">
-                        <input
-                          type="checkbox"
-                          id={`building-${activeBuilding}-canopyAddColumns-${canopyIndex}`}
-                          name={`building-${activeBuilding}-canopyAddColumns-${canopyIndex}`}
-                          checked={canopy.addColumns}
-                          onChange={(e) =>
-                            handleCanopyChange(
-                              activeBuilding,
-                              canopyIndex,
-                              'addColumns',
-                              e.target.checked
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeCanopy !== canopyIndex) {
-                              setActiveCanopy(canopyIndex);
-                            }
-                          }}
-                        />
-                        <label
-                          htmlFor={`building-${activeBuilding}-canopyAddColumns-${canopyIndex}`}
-                        >
-                          Add Columns
-                        </label>
-                      </div>
-                      <button
-                        onClick={() =>
-                          removeCanopy(activeBuilding, canopyIndex)
-                        }
-                        className="iconReject"
-                      >
-                        <FontAwesomeIcon icon={faTrash} />
-                      </button>
-                      {!isDesktop && (
-                        <>
-                          <div></div>
-                          <div className="divider span2"></div>
-                        </>
-                      )}
-                    </Fragment>
-                  )
-                )}
-                <button
-                  className="button success w5"
-                  onClick={() => addCanopy(activeBuilding)}
-                >
-                  Add
-                </button>
-              </div>
-
-              <div className="divider"></div>
-              {values.buildings[activeBuilding].canopies.length > 0 && (
-                <div className="extendGrid">
-                  <div className="extGrid start">
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-canopyRoofPanels${activeCanopy}`}
-                        name={`building-${activeBuilding}-canopyRoofPanels${activeCanopy}`}
-                        value={
-                          values.buildings[activeBuilding].canopies[
-                            activeCanopy
-                          ].roofPanelType
-                        }
-                        onChange={(e) =>
-                          handleCanopyChange(
-                            activeBuilding,
-                            activeCanopy,
-                            'roofPanelType',
-                            e.target.value
-                          )
-                        }
-                        options={roofPanels}
-                        label="Roof Panels:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-canopyRoofGauge${activeCanopy}`}
-                        name={`building-${activeBuilding}-canopyRoofGauge${activeCanopy}`}
-                        value={
-                          values.buildings[activeBuilding].canopies[
-                            activeCanopy
-                          ].roofPanelGauge
-                        }
-                        onChange={(e) =>
-                          handleCanopyChange(
-                            activeBuilding,
-                            activeCanopy,
-                            'roofPanelGauge',
-                            e.target.value
-                          )
-                        }
-                        options={roofGauge}
-                        label="Gauge:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-canopyRoofFinish${activeCanopy}`}
-                        name={`building-${activeBuilding}-canopyRoofFinish${activeCanopy}`}
-                        value={
-                          values.buildings[activeBuilding].canopies[
-                            activeCanopy
-                          ].roofPanelFinish
-                        }
-                        onChange={(e) =>
-                          handleCanopyChange(
-                            activeBuilding,
-                            activeCanopy,
-                            'roofPanelFinish',
-                            e.target.value
-                          )
-                        }
-                        options={roofFinish}
-                        label="Finish:"
-                      />
-                    </div>
-                    {selectedCanopyRoofPanel &&
-                      selectedCanopyRoofPanel.image && (
-                        <Image
-                          alt={`${selectedCanopyRoofPanel.label}`}
-                          src={selectedCanopyRoofPanel.image}
-                          className="panelImage"
-                        />
-                      )}
-                  </div>
-                  <div className="extGrid start">
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-canopySoffitPanels${activeCanopy}`}
-                        name={`building-${activeBuilding}-canopySoffitPanels${activeCanopy}`}
-                        value={
-                          values.buildings[activeBuilding].canopies[
-                            activeCanopy
-                          ].soffitPanelType
-                        }
-                        onChange={(e) =>
-                          handleCanopyChange(
-                            activeBuilding,
-                            activeCanopy,
-                            'soffitPanelType',
-                            e.target.value
-                          )
-                        }
-                        options={soffitPanels}
-                        label="Soffit Panels:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-canopySoffitGauge${activeCanopy}`}
-                        name={`building-${activeBuilding}-canopySoffitGauge${activeCanopy}`}
-                        value={
-                          values.buildings[activeBuilding].canopies[
-                            activeCanopy
-                          ].soffitPanelGauge
-                        }
-                        onChange={(e) =>
-                          handleCanopyChange(
-                            activeBuilding,
-                            activeCanopy,
-                            'soffitPanelGauge',
-                            e.target.value
-                          )
-                        }
-                        options={soffitGauge}
-                        label="Gauge:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-canopySoffitFinish${activeCanopy}`}
-                        name={`building-${activeBuilding}-canopySoffitFinish${activeCanopy}`}
-                        value={
-                          values.buildings[activeBuilding].canopies[
-                            activeCanopy
-                          ].soffitPanelFinish
-                        }
-                        onChange={(e) =>
-                          handleCanopyChange(
-                            activeBuilding,
-                            activeCanopy,
-                            'soffitPanelFinish',
-                            e.target.value
-                          )
-                        }
-                        options={soffitFinish}
-                        label="Finish:"
-                      />
-                    </div>
-                    {selectedCanopySoffitPanel &&
-                      selectedCanopySoffitPanel.image && (
-                        <Image
-                          alt={`${selectedCanopySoffitPanel.label}`}
-                          src={selectedCanopySoffitPanel.image}
-                          className="panelImage"
-                        />
-                      )}
-                  </div>
-                </div>
-              )}
-            </section>
-
-            <section className="card">
-              <header>
-                <h3>Facia</h3>
-              </header>
-            </section>
-
-            <section className="card">
-              <header>
-                <h3>Parapet Walls</h3>
-              </header>
-            </section>
-            <section className="card">
-              <header>
-                <h3>Bumpouts</h3>
-              </header>
-            </section>
+            <BuildingExtensions
+              values={values}
+              activeBuilding={activeBuilding}
+              handleNestedChange={handleNestedChange}
+              handleCanopyChange={handleCanopyChange}
+              setValues={setValues}
+              isDesktop={isDesktop}
+            />
           </>
         )}
         {/* Building Partitions Page */}
         {activeCard == 'bldg-partitions' && (
           <>
-            <section className="card start">
-              <header>
-                <h3>Partition Walls</h3>
-              </header>
-              <div className="tableGrid">
-                {values.buildings[activeBuilding].partitions.map(
-                  (partition, partitionIndex) => (
-                    <Fragment
-                      key={`building-${activeBuilding}-partition-${partitionIndex}`}
-                    >
-                      <div className="cardInput">
-                        <ReusableSelect
-                          id={`building-${activeBuilding}-partitionWall-${partitionIndex}`}
-                          name={`building-${activeBuilding}-partitionWall-${partitionIndex}`}
-                          value={partition.orientation}
-                          onChange={(e) =>
-                            handlePartitionChange(
-                              activeBuilding,
-                              partitionIndex,
-                              'orientation',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activePartition !== partitionIndex) {
-                              setActivePartition(partitionIndex);
-                            }
-                          }}
-                          options={orientations}
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-partitionStart-${partitionIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-partitionStart-${partitionIndex}`}
-                          name={`building-${activeBuilding}-partitionStart-${partitionIndex}`}
-                          value={partition.start}
-                          onChange={(e) =>
-                            handlePartitionChange(
-                              activeBuilding,
-                              partitionIndex,
-                              'start',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activePartition !== partitionIndex) {
-                              setActivePartition(partitionIndex);
-                            }
-                          }}
-                          placeholder="Feet"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-partitionEnd-${partitionIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-partitionEnd-${partitionIndex}`}
-                          name={`building-${activeBuilding}-partitionEnd-${partitionIndex}`}
-                          value={partition.end}
-                          onChange={(e) =>
-                            handlePartitionChange(
-                              activeBuilding,
-                              partitionIndex,
-                              'end',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activePartition !== partitionIndex) {
-                              setActivePartition(partitionIndex);
-                            }
-                          }}
-                          placeholder="Feet"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-partitionOffset-${partitionIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-partitionOffset-${partitionIndex}`}
-                          name={`building-${activeBuilding}-partitionOffset-${partitionIndex}`}
-                          value={partition.offset}
-                          onChange={(e) =>
-                            handlePartitionChange(
-                              activeBuilding,
-                              partitionIndex,
-                              'offset',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activePartition !== partitionIndex) {
-                              setActivePartition(partitionIndex);
-                            }
-                          }}
-                          placeholder="Feet"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-partitionHeight-${partitionIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-partitionHeight-${partitionIndex}`}
-                          name={`building-${activeBuilding}-partitionHeight-${partitionIndex}`}
-                          value={partition.height}
-                          onChange={(e) =>
-                            handlePartitionChange(
-                              activeBuilding,
-                              partitionIndex,
-                              'height',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activePartition !== partitionIndex) {
-                              setActivePartition(partitionIndex);
-                            }
-                          }}
-                          placeholder="Bay #"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-partitionBaySpacing-${partitionIndex}`}
-                        ></label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-partitionBaySpacing-${partitionIndex}`}
-                          name={`building-${activeBuilding}-partitionBaySpacing-${partitionIndex}`}
-                          value={partition.baySpacing}
-                          onChange={(e) =>
-                            handlePartitionChange(
-                              activeBuilding,
-                              partitionIndex,
-                              'baySpacing',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activePartition !== partitionIndex) {
-                              setActivePartition(partitionIndex);
-                            }
-                          }}
-                          placeholder="Separate Bays with Space"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <ReusableSelect
-                          id={`building-${activeBuilding}-partitionInsulation-${partitionIndex}`}
-                          name={`building-${activeBuilding}-partitionInsulation-${partitionIndex}`}
-                          value={partition.insulation}
-                          onChange={(e) =>
-                            handlePartitionChange(
-                              activeBuilding,
-                              partitionIndex,
-                              'insulation',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activePartition !== partitionIndex) {
-                              setActivePartition(partitionIndex);
-                            }
-                          }}
-                          options={wallInsulation}
-                        />
-                      </div>
-                      <button
-                        onClick={() =>
-                          removePartition(activeBuilding, partitionIndex)
-                        }
-                        className="iconReject"
-                      >
-                        <FontAwesomeIcon icon={faTrash} />
-                      </button>
-                      {!isDesktop && (
-                        <>
-                          <div></div>
-                          <div className="divider span2"></div>
-                        </>
-                      )}
-                    </Fragment>
-                  )
-                )}
-                <button
-                  className="button success w5"
-                  onClick={() => addPartition(activeBuilding)}
-                >
-                  Add
-                </button>
-              </div>
-
-              <div className="divider"></div>
-              {values.buildings[activeBuilding].partitions.length > 0 && (
-                <div className="extendGrid">
-                  <div className="extGrid start">
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-partitionLeftPanels${activePartition}`}
-                        name={`building-${activeBuilding}-partitionLeftPanels${activePartition}`}
-                        value={
-                          values.buildings[activeBuilding].partitions[
-                            activePartition
-                          ].leftPanelType
-                        }
-                        onChange={(e) =>
-                          handlePartitionChange(
-                            activeBuilding,
-                            activePartition,
-                            'leftPanelType',
-                            e.target.value
-                          )
-                        }
-                        options={wallPanels}
-                        label="Left Panels:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-partitionLeftGauge${activePartition}`}
-                        name={`building-${activeBuilding}-partitionLeftGauge${activePartition}`}
-                        value={
-                          values.buildings[activeBuilding].partitions[
-                            activePartition
-                          ].leftPanelGauge
-                        }
-                        onChange={(e) =>
-                          handlePartitionChange(
-                            activeBuilding,
-                            activePartition,
-                            'leftPanelGauge',
-                            e.target.value
-                          )
-                        }
-                        options={wallGauge}
-                        label="Gauge:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-partitionLeftFinish${activePartition}`}
-                        name={`building-${activeBuilding}-partitionLeftFinish${activePartition}`}
-                        value={
-                          values.buildings[activeBuilding].partitions[
-                            activePartition
-                          ].leftPanelFinish
-                        }
-                        onChange={(e) =>
-                          handlePartitionChange(
-                            activeBuilding,
-                            activePartition,
-                            'leftPanelFinish',
-                            e.target.value
-                          )
-                        }
-                        options={wallFinish}
-                        label="Finish:"
-                      />
-                    </div>
-                    {selectedPartitionLeftPanel &&
-                      selectedPartitionLeftPanel.image && (
-                        <Image
-                          alt={`${selectedPartitionLeftPanel.label}`}
-                          src={selectedPartitionLeftPanel.image}
-                          className="panelImage"
-                        />
-                      )}
-                  </div>
-                  <div className="extGrid start">
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-partitionRightPanels${activePartition}`}
-                        name={`building-${activeBuilding}-partitionRightPanels${activePartition}`}
-                        value={
-                          values.buildings[activeBuilding].partitions[
-                            activePartition
-                          ].rightPanelType
-                        }
-                        onChange={(e) =>
-                          handlePartitionChange(
-                            activeBuilding,
-                            activePartition,
-                            'rightPanelType',
-                            e.target.value
-                          )
-                        }
-                        options={wallPanels}
-                        label="Right Panels:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-partitionRightGauge${activePartition}`}
-                        name={`building-${activeBuilding}-partitionRightGauge${activePartition}`}
-                        value={
-                          values.buildings[activeBuilding].partitions[
-                            activePartition
-                          ].rightPanelGauge
-                        }
-                        onChange={(e) =>
-                          handlePartitionChange(
-                            activeBuilding,
-                            activePartition,
-                            'rightPanelGauge',
-                            e.target.value
-                          )
-                        }
-                        options={wallGauge}
-                        label="Gauge:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-partitionRightFinish${activePartition}`}
-                        name={`building-${activeBuilding}-partitionRightFinish${activePartition}`}
-                        value={
-                          values.buildings[activeBuilding].partitions[
-                            activePartition
-                          ].rightPanelFinish
-                        }
-                        onChange={(e) =>
-                          handlePartitionChange(
-                            activeBuilding,
-                            activePartition,
-                            'rightPanelFinish',
-                            e.target.value
-                          )
-                        }
-                        options={wallFinish}
-                        label="Finish:"
-                      />
-                    </div>
-                    {selectedPartitionRightPanel &&
-                      selectedPartitionRightPanel.image && (
-                        <Image
-                          alt={`${selectedPartitionRightPanel.label}`}
-                          src={selectedPartitionRightPanel.image}
-                          className="panelImage"
-                        />
-                      )}
-                  </div>
-                </div>
-              )}
-            </section>
+            <BuildingPartitions
+              values={values}
+              activeBuilding={activeBuilding}
+              handlePartitionChange={handlePartitionChange}
+              setValues={setValues}
+              isDesktop={isDesktop}
+            />
           </>
         )}
         {/* Building Options Page */}
         {activeCard == 'bldg-options' && (
           <>
-            <section className="card start">
-              <header>
-                <h3>Liner Panels</h3>
-              </header>
-              <div className="linerGrid">
-                {values.buildings[activeBuilding].linerPanels.map(
-                  (linerPanel, linerPanelIndex) => (
-                    <Fragment
-                      key={`building-${activeBuilding}-linerPanel-${linerPanelIndex}`}
-                    >
-                      <div className="cardInput">
-                        <ReusableSelect
-                          id={`building-${activeBuilding}-linerPanelWall-${linerPanelIndex}`}
-                          name={`building-${activeBuilding}-linerPanelWall-${linerPanelIndex}`}
-                          value={linerPanel.wall}
-                          onChange={(e) =>
-                            handleLinerPanelChange(
-                              activeBuilding,
-                              linerPanelIndex,
-                              'wall',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeLinerPanel !== linerPanelIndex) {
-                              setActiveLinerPanel(linerPanelIndex);
-                            }
-                          }}
-                          options={walls}
-                          label="Wall"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-linerPanelStart-${linerPanelIndex}`}
-                        >
-                          Start (Left to Right)
-                        </label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-linerPanelStart-${linerPanelIndex}`}
-                          name={`building-${activeBuilding}-linerPanelStart-${linerPanelIndex}`}
-                          value={linerPanel.start}
-                          onChange={(e) =>
-                            handleLinerPanelChange(
-                              activeBuilding,
-                              linerPanelIndex,
-                              'start',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeLinerPanel !== linerPanelIndex) {
-                              setActiveLinerPanel(linerPanelIndex);
-                            }
-                          }}
-                          placeholder="Feet"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-linerPanelEnd-${linerPanelIndex}`}
-                        >
-                          End (Left to Right)
-                        </label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-linerPanelEnd-${linerPanelIndex}`}
-                          name={`building-${activeBuilding}-linerPanelEnd-${linerPanelIndex}`}
-                          value={linerPanel.end}
-                          onChange={(e) =>
-                            handleLinerPanelChange(
-                              activeBuilding,
-                              linerPanelIndex,
-                              'end',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activeLinerPanel !== linerPanelIndex) {
-                              setActiveLinerPanel(linerPanelIndex);
-                            }
-                          }}
-                          placeholder="Feet"
-                        />
-                      </div>
-                      <div className="cardInput">
-                        <label
-                          htmlFor={`building-${activeBuilding}-linerPanelHeight-${linerPanelIndex}`}
-                        >
-                          Height
-                        </label>
-                        <input
-                          type="text"
-                          id={`building-${activeBuilding}-linerPanelHeight-${linerPanelIndex}`}
-                          name={`building-${activeBuilding}-linerPanelHeight-${linerPanelIndex}`}
-                          value={linerPanel.height}
-                          onChange={(e) =>
-                            handleLinerPanelChange(
-                              activeBuilding,
-                              linerPanelIndex,
-                              'height',
-                              e.target.value
-                            )
-                          }
-                          onFocus={() => {
-                            if (activePartition !== linerPanelIndex) {
-                              setActivePartition(linerPanelIndex);
-                            }
-                          }}
-                          placeholder="Feet"
-                        />
-                      </div>
-                      {!isDesktop && (
-                        <>
-                          <div></div>
-                        </>
-                      )}
-                      <button
-                        onClick={() =>
-                          removeLinerPanel(activeBuilding, linerPanelIndex)
-                        }
-                        className="iconReject"
-                      >
-                        <FontAwesomeIcon icon={faTrash} />
-                      </button>
-                      {!isDesktop && (
-                        <>
-                          <div className="divider span2"></div>
-                        </>
-                      )}
-                    </Fragment>
-                  )
-                )}
-                <button
-                  className="button success w5"
-                  onClick={() => addLinerPanel(activeBuilding)}
-                >
-                  Add
-                </button>
-              </div>
-
-              <div className="divider"></div>
-              {values.buildings[activeBuilding].linerPanels.length > 0 && (
-                <div className="extendGrid">
-                  <div className="extGrid start"></div>
-                  <div className="extGrid start">
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-linerPanelType${activeLinerPanel}`}
-                        name={`building-${activeBuilding}-linerPanelType${activeLinerPanel}`}
-                        value={
-                          values.buildings[activeBuilding].linerPanels[
-                            activeLinerPanel
-                          ].panelType
-                        }
-                        onChange={(e) =>
-                          handleLinerPanelChange(
-                            activeBuilding,
-                            activeLinerPanel,
-                            'panelType',
-                            e.target.value
-                          )
-                        }
-                        options={wallPanels}
-                        label="Liner Panels:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-linerPanelGauge${activeLinerPanel}`}
-                        name={`building-${activeBuilding}-linerPanelGauge${activeLinerPanel}`}
-                        value={
-                          values.buildings[activeBuilding].linerPanels[
-                            activeLinerPanel
-                          ].panelGauge
-                        }
-                        onChange={(e) =>
-                          handleLinerPanelChange(
-                            activeBuilding,
-                            activeLinerPanel,
-                            'panelGauge',
-                            e.target.value
-                          )
-                        }
-                        options={wallGauge}
-                        label="Gauge:"
-                      />
-                    </div>
-                    <div className="cardInput">
-                      <ReusableSelect
-                        id={`building-${activeBuilding}-linerPanelFinish${activeLinerPanel}`}
-                        name={`building-${activeBuilding}-linerPanelFinish${activeLinerPanel}`}
-                        value={
-                          values.buildings[activeBuilding].linerPanels[
-                            activeLinerPanel
-                          ].panelFinish
-                        }
-                        onChange={(e) =>
-                          handleLinerPanelChange(
-                            activeBuilding,
-                            activeLinerPanel,
-                            'panelFinish',
-                            e.target.value
-                          )
-                        }
-                        options={wallFinish}
-                        label="Finish:"
-                      />
-                    </div>
-                    {selectedLinerPanel && selectedLinerPanel.image && (
-                      <Image
-                        alt={`${selectedLinerPanel.label}`}
-                        src={selectedLinerPanel.image}
-                        className="panelImage"
-                      />
-                    )}
-                  </div>
-                </div>
-              )}
-            </section>
+            <BuildingOptions
+              values={values}
+              activeBuilding={activeBuilding}
+              handleNestedChange={handleNestedChange}
+              handleLinerPanelChange={handleLinerPanelChange}
+              handleWainscotChange={handleWainscotChange}
+              handlePartialWallChange={handlePartialWallChange}
+              handleWallSkirtChange={handleWallSkirtChange}
+              setValues={setValues}
+              isDesktop={isDesktop}
+            />
           </>
         )}
         {activeCard == 'bldg-cranes' && <section></section>}
@@ -1970,6 +577,7 @@ export default function ClientQuote({ session }) {
             </section>
           )}
       </form>
+      {/* Carousel Navigation */}
       {!isDesktop && (
         <nav className={styles.carouselNav}>
           <button className={styles.navArrow} onClick={handlePrev}>
