+++ src/app/(protected)/quote/_initialState.js
@@ -120,6 +120,27 @@ export const initialState = {
       canopies: [],
       partitions: [],
       linerPanels: [],
+      wainscots: [],
+      partialWalls: [],
+      wallSkirts: [],
+      fswPolySize: '3',
+      fswPolyColor: 'clear',
+      fswPolyQty: '',
+      bswPolySize: '3',
+      bswPolyColor: 'clear',
+      bswPolyQty: '',
+      lewPolySize: '3',
+      lewPolyColor: 'clear',
+      lewPolyQty: '',
+      rewPolySize: '3',
+      rewPolyColor: 'clear',
+      rewPolyQty: '',
+      backRoofPolySize: '10',
+      backRoofPolyColor: 'clear',
+      backRoofPolyQty: '',
+      frontRoofPolySize: '10',
+      frontRoofPolyColor: 'clear',
+      frontRoofPolyQty: '',
     },
   ],
 };
