+++ version.js
@@ -1,1 +1,1 @@
-export const APP_VERSION = '0.1.2';
+export const APP_VERSION = '0.1.4';
